Compile with Clang -Wshorten-64-to-32 warning flag
Unfortunately, this does not warn about the conversion from
'const uint64_t' to 'int' that MSVC warns about, which is fixed in
https://github.com/AOMediaCodec/libavif/pull/1036.

+++ tests/gtest/avifincrtest_helpers.cc
@@ -100,8 +100,8 @@ uint32_t GetMinDecodedRowCount(uint32_t height, uint32_t cell_height,
     byte_count -= byte_count / 2;
   }
   // Linearly map the input availability ratio to the decoded row ratio.
-  const uint32_t min_decoded_cell_row_count =
-      (height / cell_height) * available_byte_count / byte_count;
+  const uint32_t min_decoded_cell_row_count = static_cast<uint32_t>(
+      (height / cell_height) * available_byte_count / byte_count);
   const uint32_t min_decoded_px_row_count =
       min_decoded_cell_row_count * cell_height;
   // One cell is the incremental decoding granularity.
